+++ TravelAgencyInfrastructure/Controllers/CountriesController.cs
@@ -1,9 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
+﻿// Файл: TravelAgencyInfrastructure/Controllers/CountriesController.cs
+using System;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using TravelAgencyDomain.Model;
 using TravelAgencyInfrastructure;
@@ -28,18 +27,9 @@ namespace TravelAgencyInfrastructure.Controllers
         // GET: Countries/Details/5
         public async Task<IActionResult> Details(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var country = await _context.Countries
-                .FirstOrDefaultAsync(m => m.CountryId == id);
-            if (country == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
+            var country = await _context.Countries.FirstOrDefaultAsync(m => m.CountryId == id);
+            if (country == null) return NotFound();
             return View(country);
         }
 
@@ -50,12 +40,16 @@ namespace TravelAgencyInfrastructure.Controllers
         }
 
         // POST: Countries/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("CountryId,CountryName,VisaRequired,Continent")] Country country)
+        public async Task<IActionResult> Create([Bind("CountryName,VisaRequired,Continent")] Country country)
         {
+            // Перевірка на унікальність назви країни (без урахування регістру)
+            if (await _context.Countries.AnyAsync(c => c.CountryName.ToLower() == country.CountryName.ToLower()))
+            {
+                ModelState.AddModelError("CountryName", "Країна з такою назвою вже існує.");
+            }
+
             if (ModelState.IsValid)
             {
                 _context.Add(country);
@@ -68,29 +62,23 @@ namespace TravelAgencyInfrastructure.Controllers
         // GET: Countries/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
             var country = await _context.Countries.FindAsync(id);
-            if (country == null)
-            {
-                return NotFound();
-            }
+            if (country == null) return NotFound();
             return View(country);
         }
 
         // POST: Countries/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("CountryId,CountryName,VisaRequired,Continent")] Country country)
         {
-            if (id != country.CountryId)
+            if (id != country.CountryId) return NotFound();
+
+            // Перевірка на унікальність назви країни (без урахування регістру), окрім поточного запису
+            if (await _context.Countries.AnyAsync(c => c.CountryId != country.CountryId && c.CountryName.ToLower() == country.CountryName.ToLower()))
             {
-                return NotFound();
+                ModelState.AddModelError("CountryName", "Країна з такою назвою вже існує.");
             }
 
             if (ModelState.IsValid)
@@ -102,14 +90,8 @@ namespace TravelAgencyInfrastructure.Controllers
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!CountryExists(country.CountryId))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
+                    if (!CountryExists(country.CountryId)) return NotFound();
+                    else throw;
                 }
                 return RedirectToAction(nameof(Index));
             }
@@ -119,18 +101,9 @@ namespace TravelAgencyInfrastructure.Controllers
         // GET: Countries/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var country = await _context.Countries
-                .FirstOrDefaultAsync(m => m.CountryId == id);
-            if (country == null)
-            {
-                return NotFound();
-            }
-
+            if (id == null) return NotFound();
+            var country = await _context.Countries.FirstOrDefaultAsync(m => m.CountryId == id);
+            if (country == null) return NotFound();
             return View(country);
         }
 
@@ -139,19 +112,31 @@ namespace TravelAgencyInfrastructure.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (_context.Countries == null)
+            {
+                return Problem("Entity set 'TravelAgencyDbContext.Countries' is null.");
+            }
             var country = await _context.Countries.FindAsync(id);
             if (country != null)
             {
-                _context.Countries.Remove(country);
-            }
+                bool hasRelatedTours = await _context.Tours.AnyAsync(t => t.CountryId == id);
+                bool hasRelatedHotels = await _context.Hotels.AnyAsync(h => h.CountryId == id);
 
-            await _context.SaveChangesAsync();
+                if (hasRelatedTours || hasRelatedHotels)
+                {
+                    ModelState.AddModelError(string.Empty, "Неможливо видалити країну. Вона має пов'язані тури або готелі. Спочатку видаліть або перепризначте їх.");
+                    var countryToDeleteView = await _context.Countries.FirstOrDefaultAsync(m => m.CountryId == id); // Повторно завантажуємо для View
+                    return View("Delete", countryToDeleteView);
+                }
+                _context.Countries.Remove(country);
+                await _context.SaveChangesAsync();
+            }
             return RedirectToAction(nameof(Index));
         }
 
         private bool CountryExists(int id)
         {
-            return _context.Countries.Any(e => e.CountryId == id);
+            return (_context.Countries?.Any(e => e.CountryId == id)).GetValueOrDefault();
         }
     }
 }
